avoid GetComponent() calls in patched code
I've done this for my other mods, FastTrack logs suggest
GetComponent() may be somewhat expensive when called often,
for this mod I don't actually see it in the logs, but let's
do it anyway. This mod's UpdateLogicCircuit() calls mapping
the other way around should be fine, as they are only called
as a result of UI input.